conv

+++ assignment_3/src/Flatten.py
@@ -8,11 +8,11 @@ class Flatten:
 		self.gradInput = None
 	
 	def forward(self, inp):
-		self.output = inp.reshape((inp.shape[0], -1))
+		self.output = inp.reshape((inp.shape[0], -1)).double()
 		return self.output
 
 	def backward(self, inp, gradOutput, lr=None):
-		self.gradInput = gradOutput.reshape(inp.shape)
+		self.gradInput = gradOutput.reshape(inp.shape).double()
 		return self.gradInput
 
 	def set_device(self, device):
